+++ test/models/user.rb
@@ -8,6 +8,8 @@ class User
 
   validates_confirmation_of :password
 
+  validates_format_of :email, with: /NOSPAM/
+
   validates_length_of :name, minimum: 10, maximum: 100
   validates_length_of :password, is: 8
   validates_size_of :lastname, maximum: 100