+++ CHANGELOG.md
@@ -15,3 +15,4 @@
 + Add `in` alias to `ValidateLengthMatcher#within`.
 + Add `is` option to `ValidateLengthMatcher`.
 + Add `is_equal_to` alias to `ValidateLengthMatcher#is`.
++ Add `validate_format_of` matcher.